Bumped version number to 427.0.0
We are changing the way we bump version numbers. In the past, we would
do dev work on 426.0.0-dev.1 and cut the final release onto 427.0.0. We
don't want to continue doing this. So we are skipping a version in order
to bump releases as finals. This means 426.0.0-dev.1 will become
426.0.0. Git tags and branches will also follow this pattern.

Signed-off-by: Jonathan Qiwen Tsang <jtsang@pivotal.io>