Merge release verison into provided manifest
As part of the story to upgrade pcf-tile-pipeline, we are required to
inject our release version into the manifest. In the long run, it is
expected that we will not perform BOSH operations within tests, but be
given a BOSH environment to perform tests in instead. This
responsibility will live with pcf-tile-pipeline. At the moment we simply
deploy the latest version of cf-redis available to the director. We need
to avoid this by injecting the release version directly into the
manifest.

[#131716239]

Signed-off-by: Jonathan Qiwen Tsang <jtsang@pivotal.io>