Specify stdout_filename and stderr_filename
We have not been configuring our redis backup logs correctly. Our code
requires a file to be present in order for it to log stdout and stderr.
We had coded this into the ctl scripts of the broker and the redis-agent
on dedicated nodes, but we forgot to insert these properties into the
spec file of the jobs. Adding this back in has fixed logging for redis
backups.

[#134155847]