Fix orphaned sleeps and ignored pkill error 1
- Using exec sleep instead of simply sleep means that sleep processes
run as children of the invoking one. Therefore they die with it and
don't remain around as orphaned processes.
- pkill returns 1 if there is nothing to kill; this is fine (it means
there were no redis-server in the first place) so we ignore
that exit code in the process destroyer stop script. Any other non-zero
exit code from pkill will cause the stop script to fail.

[#137029011]

Signed-off-by: Maria Ntalla <mntalla@pivotal.io>