Refactor unnecessary waiting for broker
We used to wait for the broker to come back to life before performing
any deprovisions. Now we don't need to because we don't bring down the
broker during the story. It was still a valid test to keep around for
the failure cases so we moved it from the after block to the expectation
block.

[#135936747]

Signed-off-by: Jonathan Qiwen Tsang <jtsang@pivotal.io>