Remove exec from loop in syslog-configurator
The exec was placed in the ctl script to avoid the creation of orphaned
sleep processes. This unfortunately led the infinite loop to break after
the first cycle. Why this is the case is still unknown, but for now we
will revert this change. The bug would cause syslog-configurator to stop,
which would be revived by monit, which would lead redis to also be
restart roughly every 10000 seconds.

[#141017007]

Signed-off-by: Maria Ntalla <mntalla@pivotal.io>