Run tests against local bosh lite
- Updated readme to call out that almost all the variables are required
test dependencies
- Added execute command to bosh2 cli wrapper. This was required as the
previous tests required explicit version setting. In the test setup, we
use latest so bosh recreate has a diff. Switching to bosh2 allows me to
not modify hula to pass in the '-n' flag. It's also ultimately the
direction to which we're switching.
- Renamed the class Instances to Deployment as it acts on a deployment
rather than just the instances

[#159233524]